website/requirements/requirements.txt
Django==1.6
PIL==1.1.7
Pillow==2.2.1
South==0.8.4
Werkzeug==0.9.4
django-appconf==0.6
django-classy-tags==0.4
django-extensions==1.2.5
django-mptt==0.6.0
django-polymorphic==0.5
django-sekizai==0.7
django-treeadmin==0.4.1
easy-thumbnails==1.4
gunicorn==18.0
html5lib==1.0b1
httplib2==0.8
six==1.4.1
wsgiref==0.1.2